Looks good enough for now
It's inefficient, but it looks like it gets the job done. Massive
changes for SendRecvUDPwithParity.  I removed my previous debug changes
to TestSendRecv and TestCases. Using parity can be modified to change
the size of the packets (I currently send 508 bytes) and also the number
of packets sent before creating parity packet. Right now if the size of
the data block is below the recommended size for packets then no parity
packets are created. Therefore sending < 508 data blocks results in
regular UDP with no parity.